v1.1: Solved the issue of having just one app.
v1.1: as stated in developer.android.com/auto, «Apps that work with Android Auto consoles run on a connected device, such as a phone or tablet. » and this raised the issue of having just one application for the communication with the system.
Added [A16], [A18]. 
Major changes to paragraphs 5.1, 5.2, 7.5, use case 7.2.4.